refactor: CollectionView 관련 분리
